All basic conversions (not IEEE) in vue components
