--- src/views/conversions.html
@@ -1,120 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-
-    <meta charset="UTF-8">
-    <title>Sovereign Logic</title>
-
-    <!-- Bootstrap core CSS -->
-    <link href="../css/bootstrap.min.css" rel="stylesheet">
-    <!-- Bootstrap theme -->
-    <link href="../css/bootstrap-theme.min.css" rel="stylesheet">
-
-    <!-- Custom styles for this template -->
-    <link href="../css/workspace.css" rel="stylesheet">
-
-    <style>
-        #output {
-            margin-top: 62px;
-        }
-        #convert-component {
-            line-height: 2;
-        }
-    </style>
-
-</head>
-
-<body>
-
-<!-- Fixed navbar -->
-<nav class="navbar navbar-inverse navbar-fixed-top">
-    <div class="container-fluid">
-        <div class="navbar-header">
-            <img class="icon-bar" src="../images/Logo.svg">
-            <a class="navbar-brand">Sovereign Logic</a>
-        </div>
-        <div id="navbar" class="navbar-collapse collapse">
-            <ul class="nav navbar-nav">
-                <li><a href="../../index.html">Home</a></li>
-                <li><a href="binaryArithmetic.html">Binary Arithmetic</a></li>
-                <li class="active"><a href="conversions.html">Conversions</a></li>
-                <li><a href="truthTables.html">Min/Maxterms</a></li>
-                <li><a href="karnaughMaps.html">Karnaugh Maps</a></li>
-                <li><a href="designProblems.html">Design Problems</a></li>
-                <li><a href="circuitDesign.html">Circuit Design</a></li>
-                <li><a href="references.html">References</a></li>
-            </ul>
-        </div>
-    </div>
-</nav>
-
-<div class="container-fluid">
-    <div class="row">
-        <div class="col-sm-4 col-md-3 sidebar">
-            <ul class="nav nav-sidebar">
-                <li class="sub-header">Select Base to Convert From</li>
-                <li id="binary_conversions"><a href="#">Binary (Base 2)</a></li>
-                <li id="octal_conversions"><a href="#">Octal (Base 8)</a></li>
-                <li id="decimal_conversions"><a href="#">Decimal (Base 10)</a></li>
-                <li id="hex_conversions"><a href="#">Hexadecimal (Base 16)</a></li>
-            </ul>
-            <ul class="nav nav-sidebar">
-                <li class="sub-header">Special Conversions</li>
-                <li id="ieee_conversions"><a href="#">IEEE Floating Point</a></li>
-            </ul>
-        </div>
-        <div class="col-sm-8 col-sm-offset-4 col-md-9 col-md-offset-3 main">
-            <h1 id="conversions-header" class="page-header"><button type="button" style="margin-right: 12px" class="btn btn-info btn-lg" data-toggle="modal" data-target="#conversionModal">
-                Instructions!
-            </button>
-                <div class="modal fade" id="conversionModal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
-                    <div class="modal-dialog" role="document">
-                        <div class="modal-content">
-                            <div class="modal-header">
-                                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span>&times;</span></button>
-                                <h4 class="modal-title" id="conversionTitle">Conversion Instructions</h4>
-                            </div>
-                            <div class="modal-body">
-                                <h3 class="sub-header">Calculator</h3>
-                                Simply enter your decimal value and hit convert! Select the desired base
-                                you wish to convert to in the drop-down box.
-                                <h3 class="sub-header">Practice Problems</h3>
-                                This feature generates a random decimal value for you to convert.
-                                Generate new random values with the Change Values button. Check your
-                                answer with the Check Answer button. If you are incorrect, the correct answer will show.
-                                Make sure you select which base you converted to in the drop-down box as you put in your answer.
-                            </div>
-                            <div class="modal-footer">
-                                <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                            </div>
-                        </div>
-                    </div>
-                </div>Conversions from Decimal (Base 10)</h1>
-            <div class="user-input-converter">
-                <h2 class="sub-header" style="margin-bottom: 20px;">Converter</h2>
-                <div id="convert-component"></div>
-                <p id="output" class="well converter-output">The Answer Will Appear Here.</p>
-            </div>
-            <div class="conversion-quiz">
-                <h2 class="sub-header" id="second_header">Practice Problems</h2>
-                <div id="convert-quiz-component"></div>
-                <p id="quiz_output" class="well converter-output">Your Answer Will Be Checked Here.</p>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- Bootstrap core JavaScript
-================================================== -->
-<!-- Placed at the end of the document so the pages load faster -->
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-<script>window.jQuery || document.write('<script src="../../assets/js/vendor/jquery.min.js"><\/script>')</script>
-<script src="../js/libs/bootstrap.min.js"></script>
-<script src="../js/libs/docs.min.js"></script>
-<script src="../js/libs/vue.min.js"></script>
-<script src="../js/conversions.js"></script>
-<script src="../js/ieeeConversions.js"></script>
-
-</body>
-
-</html>
